requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.8
Werkzeug==2.0.2
XlsxWriter==3.0.2
